Rag-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
qdrant-client==1.7.0
openai==1.3.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
markdown==3.5.1
httpx==0.25.2
better-auth>=0.0.1b11
sqlalchemy>=1.4.0